requirements.txt
confluent_kafka==2.5.0
cassandra-driver==3.29.1
pymongo==4.8.0
apache-airflow-providers-discord==3.7.1